* We have audio playback at last! The sample file was an MP3 encoded audio music track and while it did not sound pitch perfect through Small World Deluxe, it's a start! This is a demonstration that playback through OpenAL via FFmpeg can indeed be done, within C/C++ code.

+++ src/defines.hpp
@@ -1247,7 +1247,7 @@ namespace GkAudioFramework {
 
     struct GkAudioFileDecoded {
         GkAudioFileInfo properties;
-        std::vector<double> samples;
+        std::vector<float> samples;
     };
 }
 };
